+++ apm-agent-core/src/main/java/co/elastic/apm/agent/configuration/CoreConfiguration.java
@@ -254,7 +254,8 @@ public class CoreConfiguration extends ConfigurationOptionProvider {
             "\n" +
             "A few examples:\n" +
             "\n" +
-            " - `org.example.*#*`\n" +
+            " - `org.example.*` (omitting the method is possible since 1.4.0)\n" +
+            " - `org.example.*#*` (before 1.4.0, you need to specify a method matcher)\n" +
             " - `org.example.MyClass#myMethod`\n" +
             " - `org.example.MyClass#myMethod()`\n" +
             " - `org.example.MyClass#myMethod(java.lang.String)`\n" +
@@ -267,8 +268,7 @@ public class CoreConfiguration extends ConfigurationOptionProvider {
             "The more methods you match the more overhead will be caused by the agent.\n" +
             "Also note that there is a maximum amount of spans per transaction (see <<config-transaction-max-spans, `transaction_max_spans`>>).\n" +
             "\n" +
-            "NOTE: A method matcher is always required even if you are trying to trace every method in a package.\n" +
-            "This means that `org.example.*` is invalid, while `org.example.*#*` works.")
+            "Since 1.3.0")
         .buildWithDefault(Collections.<MethodMatcher>emptyList());
 
     public boolean isActive() {